pytorch/requirements.txt
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
